Allow G0/G1 arguments I and S<value>
Some GCode dialects (e.g. Marlin) use `I` on a `G0` or `G1` command to indicate that a laser should be on for this move. 
Some GCode dialects (e.g. Marlin) use `S` on a `G0` or `G1` command to indicate the speed of a spindle or intensity of a laser.